+++ contracts/interfaces/IAccountBalance.sol
@@ -11,7 +11,7 @@ interface IAccountBalance {
     /// @param amount The amount changed
     event PnlRealized(address indexed trader, address indexed baseToken, int256 amount);
     /// @dev Emit when underlying price index is set
-    event UnderlyingPriceIndexSet(uint64 indexed underlyingIndex);
+    event UnderlyingPriceIndexSet(uint256 indexed underlyingIndex);
 
     function initialize(address positioningConfigArg) external;
     /// @notice Modify trader owedRealizedPnl
