+++ contracts/interfaces/IPositioning.sol
@@ -75,7 +75,7 @@ interface IPositioning is IFundingRate {
     event FundingPeriodSet(uint256 fundingInterval);
 
     /// @dev this function is public for testing
-    function initialize(address positioningConfigArg, address vaultControllerArg, address accountBalanceArg, address matchingEngineArg, address markPriceArg, address indexPriceArg, uint64 underlyingPriceIndex, address[2] calldata liquidators) external;
+    function initialize(address positioningConfigArg, address vaultControllerArg, address accountBalanceArg, address matchingEngineArg, address markPriceArg, address indexPriceArg, uint256 underlyingPriceIndex, address[2] calldata liquidators) external;
     /// @notice Settle all markets fundingPayment to owedRealized Pnl
     /// @param trader The address of trader
     function settleAllFunding(address trader) external;
